Shuffle functions around and move them to file utils
